Ensure form validation always runs on submit.
On some of the forms there are actions which cause the form to be submitted programatically, with submit(). It turns out this does not trigger the onsubmit event, causing validation to be skipped over. There are a few ways to get around this.

- create and dispatch event explicitly - this is not supported in IE
- use jQuery's trigger() - it seems like overkill to include jQuery just for that purpose
- explicitly call form validation where needed
- click() the submit button - has support back to IE8, but it doesn't look very intuitive

At the moment I've gone with the last option.