Fix issue where some appointment statuses were not using the correct icon when signed/verified.
There was a check which did not allow the special icon versions (S/V) to be used for non-editable statuses. The intent was to avoid a case where an icon is missing for a custom status, as these only go up to 11 at the moment. I think this concern was misplaced, since the regular icons (non S/V) for custom statuses only go up to 11 as well.

It's not ideal to rely on the existance of these icons, but since status addition is a manual back-end task I think it is not unreasonable to expect that if anyone wants to go over 11, they should be responsible for adding ALL necessary icons for their new status.