Fixes for patient search.
This method of searching, which was previously used by the new UI and is now to be used by both, had several issues.

Regular expressions. Any user input that did not contain "%" was treated as a regular expression. This was bad for several reasons. It depended on the user to input a valid regex expression and produced errors otherwise. It neglected the existance of the other sql wildcard so "%" worked as expected but "_" was treated literally. And most importantly, it was not returning the correct results. For example, say you had patients Smith and Johnson. You enter 's', expecting to see Smith only, but you get both since the regex 's' matches anything that contains the letter.
It seems unlikely that our users would be interested in using regex syntax. This change goes back to supporting sql wildcards only.

The behaviour should now match the old search:
Searching by name, HIN, or chart number looks for the keyword at the start of the value.
Searching by phone or address looks for the keyword anywhere in the value.
Searching by demographic number looks for the keyword to match the value exactly.

With these two changes:
Searching by date of birth will now look to match the keyword exactly. Previously a search like '19-1-2' would match entries such as '1982-10-21'.
Searching with wildcards means you are taking charge of the exact expression to be used. For example, a name search 'j' will return results for 'j%' however a search 'j%on' will return results that match exactly that (not 'j%on%' as the old search would have done). Using this option always wins over the default, ie you can do wildcard searches on demographic number and date.