Fix for empty fax file errors: on unsuccessful wkhtmltopdf try again up to a number of attempts then give up.
The problem was that for eForms a temporary pdf file is created and then populated by wkhtmltopdf. However if wkhtmltopdf hangs for some reason (perhaps connection issues), it is terminated after some time (currently 40s). The temp file is then still empty but it attempts to fax out anyway, which obviously doesn't work.

I don't think we can do anything to guarantee a successful PDF conversion, so if it didn't work just let the user know that the fax failed.