Fix for error on attempting to delete measurement type.
This was fixed in the open source OSCAR project earlier this year by removing the PreRemove and PreUpdate entity listeners on measurement types. I've only removed the PreRemove here as that is what was causing the error and I don't know if we want to allow updates for measurement types - I don't think that's exposed to the user at the moment.

This error was also hiding another issue where the measurement group entries were not being deleted properly either.