Remove patch turning Juno UI off and simplify new init patch.
provider_no 999998 is still having property table populated via initial
DB create, so I'm not modifying that. We do need the INSERT IGNORE as a
result to run this without error.

Also added the missing oscarResources property to other files, and added
a constant for the "yes"/"no" values in the property table to
UserProperty.