Escape quotes where required, remove otherwise
Single quotes needed to be escaped on the ManageContacts page.

For contactSearch, there were issues when attempting to click on a
contact entry containing double quotes anywhere in the name. The quotes
seem to be only for display purposes, so I decided to strip them out and
escape the name. The quotes would be properly re-displayed when
re-viewing the contacts page after linking a contact to a demographic.