Redo some of appointment status modifying to loosen super admin reqs
Super admin only required now for changing order of statuses and adding
new ones. Everything else should be accessible to an admin user.

Changing order of statuses is temporarily super admin only until we can
confirm that it works as we expect; after it's confirmed, we should
remove the super admin bit and only allow reordering of statuses that
are flagged as editable in the database.

For all users, when a status is flagged as "not editable" in the
database you can't disable it or change its name.