Prevent non super admin users from messing with super admin record
Also sneaking in a change around K2A - if a provider didn't have admin
role, an exception would be thrown internally. Clogged up logs a lot.

Other change being sneaked in is around getting ProviderPreference
initialized - if for whatever reason we had to build a new object to
persist, it wouldn't actually persist as there was no primary key to
autogenerate. Fix is to use passed in provider number that we were using
to pull record out to begin with.