Use criteria search instead of JSQL query
Also add capability for criteria search to filter based off a version
code. Added inversion check for this particular case (look for items
where version code is NOT this) because of BC's good ol' baby clause,
but most of the time you would probably want an exact match were you to
supply it.