Properly handle case where consultation has no specialist
Consultations are expected to be allowed to not have a specialist. Juno
UI originally made the assumption that it was required, but classic lets
you do this.

Also moved a bunch of other old converters to our new style of
conversion since I had to make adjustments in this area.