Split out validate / build model / save for measurements
This was all being done in one gross, ugly function within the action. I
would have liked to make it so that none of the measurements save, but
the error display logic here is so tightly coupled to the classic UI
that it's not feasible to do within the context of my current problem.

This should hopefully make it a lot clearer as to what this action is
doing. None of the logic has really changed (hopefully), just split up
so that each distinct procedure is actually a distinct procedure.