Move termination reason to enum
Was sitting in a properties file before and being referenced via some
very old code that loaded this properties file up.

There's some other spots on the Juno UI that need to be aware of where
this enum lives, but for now this is minimally enough to get things to
load/persist using the enum and not break all of the things.