Add basic roster status management to Juno UI
Roster status management currently can do:
- adding of new statuses
- editing of new statuses
It currently looks terrible due to lack of specific styling, and also
needs at a basic level the ability to flag a status as active or
inactive (to hide from dropdown).

Tried using a drag-and-drop reorder but it was too time consuming to get
it working 100% correct, so I'm not adding the capability to reorder
statuses for now.

Also addressed various minor bits of code review feedback.