Code review feedback
Also removing foreign key for demographic_roster ahead of future
changes; rostered provider isn't necessarily a *provider*.

We don't know what it should be quite yet, because the requirement
around this is incredibly unclear. But it can be the MRP but isn't
necessarily the MRP, and it may not be a provider at the clinic at all,
meaning that limiting ourselves to providers at the clinic is
overreaching.