Include more in demographic transfers
Attempted to do a more "proper" transfer object that better matched the
JPA demographic model but it was far too
widespread of a change to do properly within the scope of my changes.

For now, converting what is necessary to ensure the Juno UI demographic
record can still do its thing in full and not drop any inputs.