Rework prevention printing
Changes were originally made against Juno, this is pulling a working
version of it and committing it against Spring Boot.

Had to make some changes around some Gender stuff and some ClinicService
things that seem to not be up to date under Spring Boot. Would rather
just get this working and maybe later we can fix this if we're ever back
in here.