Attempt at detecting / dynamic URL switching for PPN
CareConnect doesn't actually provide any way of detecting whether you
are on their Physician's Private Network (PPN), but expect us to be able
to dynamically adjust depending on whether we are on it.

We also have no way of actually getting onto the PPN to test whether
this works, so this is all conjecture and *assumes* that if you are on
the PPN that you can make a GET request to it.