Do enrolled doctor search update from backend
Bulk of this commit is to add the new converters using our new converter
pattern, to replace the old one. Some of the toDomain converters are
largely stubs in case they ever are needed,
but my hope is that they're never needed.

Also added unit tests for the one converter that I'm actually relying on
for the enrolled doctor field.